refactor(error): split ErrorPage into reusable components for better maintainability
